% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/json_types.r
\docType{data}
\name{allowed_json_types}
\alias{allowed_json_types}
\title{Fundamental JSON types from http://json.org/, where I collapse 'true' and
'false' into 'logical'}
\format{\preformatted{ chr [1:6] "object" "array" "string" "number" "logical" "null"
}}
\usage{
allowed_json_types
}
\description{
Fundamental JSON types from http://json.org/, where I collapse 'true' and
'false' into 'logical'
}
\keyword{datasets}

