% Generated by roxygen2 (4.0.1): do not edit by hand
\name{append}
\alias{append}
\alias{append_logical}
\alias{append_number}
\alias{append_string}
\title{Append keys to a new column}
\usage{
append_string(x, column.name = "string")

append_number(x, column.name = "number")

append_logical(x, column.name = "logical")
}
\description{
Append keys to a new column
}

