% Generated by roxygen2 (4.1.0): do not edit by hand
\name{append_values}
\alias{append_values}
\alias{append_values_logical}
\alias{append_values_number}
\alias{append_values_string}
\title{Append keys to a new column}
\usage{
append_values_string(x, column.name = type, force = TRUE)

append_values_number(x, column.name = type, force = TRUE)

append_values_logical(x, column.name = type, force = TRUE)
}
\arguments{
\item{x}{a tbl_json object}

\item{column.name}{the column.name to append the values into the data.frame
under}

\item{force}{parameter that determines if the variable type should be computed or not
if force is FALSE, then the function may take more memory}
}
\description{
Append keys to a new column
}

