% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/append_values.r
\name{append_values_factory}
\alias{append_values_factory}
\title{Creates the append_values_* functions}
\usage{
append_values_factory(type, na_value, blank_value, force = TRUE)
}
\arguments{
\item{type}{the JSON type that will be appended}

\item{na_value}{the NA value that will be used if the type is incorrect}

\item{force}{parameter that determines if the type should be determined or not
if force is FALSE, then the function takes more memory}
}
\description{
Creates the append_values_* functions
}

