% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/append_values.r
\name{append_values_type}
\alias{append_values_type}
\title{get list of values from json}
\usage{
append_values_type(json, type, na_value)
}
\arguments{
\item{json}{extracted using attributes}

\item{type}{input type (numeric, string, etc)}

\item{na_value}{default null value}
}
\description{
get list of values from json
}

