% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_json.R
\name{as.character.tbl_json}
\alias{as.character.tbl_json}
\title{Convert the JSON in an tbl_json object back to a JSON string}
\usage{
\method{as.character}{tbl_json}(x, ...)
}
\arguments{
\item{x}{a tbl_json object}

\item{...}{not used (\code{\link[purrr]{map_chr}} used instead)}
}
\value{
a character vector of formatted JSON
}
\description{
Convert the JSON in an tbl_json object back to a JSON string
}
