% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_json.R
\name{as_tibble.tbl_json}
\alias{as_tibble.tbl_json}
\alias{as_data_frame.tbl_json}
\title{Convert a tbl_json back to a tbl_df}
\usage{
\method{as_tibble}{tbl_json}(x, ...)

as_data_frame.tbl_json(x, ...)
}
\arguments{
\item{x}{a tbl_json object}

\item{...}{additional parameters}
}
\value{
a tbl_df object (with no tbl_json component)
}
\description{
Drops the JSON attribute and the tbl_json class, so that
we are back to a pure tbl_df.  Useful for some internals.  Also useful
when you are done processing the JSON portion of your data and are
ready to move on to other tools.
}
\details{
Note that as.tbl calls tbl_df under the covers, which in turn
calls as_tibble.  As a result, this should take care of all cases.
}
