% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{companies}
\alias{companies}
\title{Startup company information for 1,000 companies}
\format{JSON}
\usage{
companies
}
\description{
From: http://jsonstudio.com/resources/
}
\examples{
library(dplyr)

# Print the first record
companies[[1]] \%>\% writeLines

# Get the key employees data
key_employees <- companies \%>\%
  as.tbl_json \%>\%
  spread_values(
    name = jstring("name")
  ) \%>\%
  enter_object("relationships") \%>\%
  gather_array("relationship.index") \%>\%
  spread_values(
    is.past = jstring("is_past"),
    name = jstring("person", "permalink"),
    title = jstring("title")
  )

# Show the top 10 titles
key_employees \%>\%
  filter(is.past == "FALSE") \%>\%
  group_by(title) \%>\%
  tally() \%>\%
  arrange(desc(n)) \%>\%
  top_n(10)
}

