% Generated by roxygen2 (4.1.0): do not edit by hand
\name{determine_types}
\alias{determine_types}
\title{Determines the types of a list of parsed JSON}
\usage{
determine_types(json_list)
}
\arguments{
\item{json_list}{a list of parsed JSON}
}
\value{
a factor with levels json_types
}
\description{
Determines the types of a list of parsed JSON
}

