% Generated by roxygen2 (4.0.1): do not edit by hand
\name{enter_object}
\alias{enter_object}
\title{Selects an object by key and filters rows to just those with matching keys}
\usage{
enter_object(x, ...)
}
\arguments{
\item{x}{a tbl_json object}

\item{...}{path to filter}
}
\description{
Selects an object by key and filters rows to just those with matching keys
}

