% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{filter}
\alias{filter}
\title{Objects exported from other packages}
\details{
This needs to be re-exported, since `dplyr` implements a generic that 
is not a generic in the `stats` package, and `tidyjson` provides a method for that generic
}
\seealso{
http://r.789695.n4.nabble.com/R-CMD-check-warning-with-S3-method-td4692255.html
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}}
}}

