% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{gather_factory}
\alias{gather_factory}
\title{Factory to create gather functions}
\usage{
gather_factory(
  default.column.name,
  default.column.empty,
  expand.fun,
  required.type
)
}
\arguments{
\item{default.column.name}{the desired name of the default added column}

\item{default.column.empty}{the value to use when the default column should
be empty because the JSON has length 0}

\item{expand.fun}{a function applied to the JSON that will expand the rows
in the tbl_df}

\item{required.type}{the json_types type that must be present in every
element of the JSON for this to succeed}
}
\description{
Factory to create gather functions
}
\keyword{internal}
