% Generated by roxygen2 (4.1.0): do not edit by hand
\name{gather_keys}
\alias{gather_keys}
\title{Gathers every key from the top level of the json and stacks them}
\usage{
gather_keys(x, column.name = "key")
}
\arguments{
\item{x}{a tbl_json whose JSON attribute should always be an object}

\item{column.name}{the name to give to the column of key names created}
}
\value{
a tbl_json with a new column (column.name) that captures the keys
  and JSON attribute of the associated value data
}
\description{
Gathers every key from the top level of the json and stacks them
}

