% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_names}
\alias{has_names}
\title{Check for Names}
\usage{
has_names(x)
}
\arguments{
\item{x}{Input object}
}
\value{
Boolean.  Indicates whether x has names
}
\description{
Checks the input object for the existence of names
}
\keyword{internal}
