% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_data_list}
\alias{is_data_list}
\title{List Check}
\usage{
is_data_list(x)
}
\arguments{
\item{x}{Input object}
}
\value{
Boolean.  Indicates whether x is a list
}
\description{
Checks whether a list is being provided
}
\keyword{internal}
