% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_json.R
\name{is_json_factory}
\alias{is_json_factory}
\title{Factory to create \code{is_json} functions}
\usage{
is_json_factory(desired.types)
}
\arguments{
\item{desired.types}{character vector of types we want to check for}
}
\value{
a function
}
\description{
Factory to create \code{is_json} functions
}
\keyword{internal}
