% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{issues}
\alias{issues}
\title{Issue data for the dplyr repo from github API}
\format{JSON}
\usage{
issues
}
\description{
Issue data for the dplyr repo from github API
}
\examples{
library(dplyr)

# Show first 2k characters of JSON
issues \%>\% substr(1, 2000) \%>\% writeLines
}

