% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.jdf}
\alias{as.jdf}
\alias{as.jdf.character}
\alias{as.jdf.jdf}
\title{Turns a character vector into a jdf object}
\usage{
as.jdf(x, ...)

\method{as.jdf}{jdf}(x, ...)

\method{as.jdf}{character}(json, ...)
}
\arguments{
\item{json}{character vector of json}
}
\description{
Turns a character vector into a jdf object
}

