% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jfilter}
\alias{jfilter}
\title{Selects an object by key and filters rows to just those with matching keys}
\usage{
jfilter(x, ...)
}
\arguments{
\item{...}{path to filter}
}
\description{
Selects an object by key and filters rows to just those with matching keys
}

