% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_values.R
\name{jfunctions}
\alias{jfunctions}
\alias{jlogical}
\alias{jnumber}
\alias{jstring}
\title{Navigates nested objects to get at keys of a specific type, to be used as
arguments to spread_values}
\usage{
jstring(..., recursive = FALSE)

jnumber(..., recursive = FALSE)

jlogical(..., recursive = FALSE)
}
\arguments{
\item{...}{the path to follow}

\item{recursive}{logical indicating whether second level and beyond objects
should be extracted.  Only works when there exists a single value in
the nested json object}
}
\value{
a function that can operate on parsed JSON data
}
\description{
Note that these functions fail if they encounter the incorrect type.
}

