% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_values.R
\name{json_factory}
\alias{json_factory}
\title{Factory that creates the j* functions below}
\usage{
json_factory(map.function)
}
\arguments{
\item{map.function}{function to map to collapse}
}
\description{
Factory that creates the j* functions below
}
\keyword{internal}
