% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_values.R
\name{json_functions}
\alias{json_functions}
\alias{json_chr}
\alias{jstring}
\alias{json_dbl}
\alias{jnumber}
\alias{json_lgl}
\alias{jlogical}
\title{Navigates nested objects to get at names of a specific type, to be used as
arguments to \code{\link{spread_values}}}
\usage{
json_chr(..., recursive = FALSE)

jstring(..., recursive = FALSE)

json_dbl(..., recursive = FALSE)

jnumber(..., recursive = FALSE)

json_lgl(..., recursive = FALSE)

jlogical(..., recursive = FALSE)
}
\arguments{
\item{...}{a quoted or unquoted sequence of strings designating the object
name sequence you wish to follow to find a value}

\item{recursive}{logical indicating whether second level and beyond objects
should be extracted.  Only works when there exists a single value in
the nested json object}
}
\value{
a function that can operate on parsed JSON data
}
\description{
Note that these functions fail if they encounter the incorrect type.
}
\seealso{
\code{\link{spread_values}} for using these functions to spread
         the values of a JSON object into new columns
}
