% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_json.R
\name{json_get}
\alias{json_get}
\title{Raw JSON}
\usage{
json_get(.data)
}
\arguments{
\item{.data}{A tbl_json object}
}
\value{
A nested list representing the JSON data
}
\description{
Retrieve the raw JSON from a tbl_json object
}
