% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_json.R
\name{json_get_column}
\alias{json_get_column}
\title{Make the JSON data a persistent column}
\usage{
json_get_column(.data, column_name = "json")
}
\arguments{
\item{.data}{A tbl_json object}

\item{column_name}{Optional. The name of the output column (either as a
string or unquoted name). Default "json"}
}
\value{
A tbl_json object with an added column containing the JSON data
}
\description{
Extract the raw JSON from a tbl_json object. Store it in a column. WARNING:
column name collisions will be overwritten
}
\examples{

tj <- as_tbl_json('{"a": "b"}')
json_get_column(tj, my_json)

}
