% Generated by roxygen2 (4.1.0): do not edit by hand
\name{json_lengths}
\alias{json_lengths}
\title{Add a column that tells the 'length' of the data in the root of the JSON}
\usage{
json_lengths(x, column.name = "length")
}
\arguments{
\item{x}{a tbl_json object}

\item{column.name}{the name to specify for the length column}
}
\value{
a tbl_json object with column.name column that tells the length
}
\description{
Add a column that tells the 'length' of the data in the root of the JSON
}

