% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_schema.R
\name{json_schema}
\alias{json_schema}
\title{Summarize a collection of JSON documents into a single schema document}
\usage{
json_schema(x)
}
\arguments{
\item{x}{a json string or a tbl_json object}
}
\description{
Summarize a collection of JSON documents into a single schema document
}

