% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_or_dots}
\alias{list_or_dots}
\title{List or Dots}
\usage{
list_or_dots(...)
}
\arguments{
\item{...}{Either a list or the `...` of a function call}
}
\value{
The input object coerced into a list for easier use
}
\description{
Handles dots or a list, coercing into a list
so that the output is easy to handle
}
