% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/path.r
\name{list_path}
\alias{list_path}
\title{Recursively access a path}
\usage{
list_path(l, path)
}
\arguments{
\item{l}{a list}

\item{path}{a path of keys to follow}
}
\description{
Recursively access a path
}

