% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\title{Create a JSON path with a minimum of typing}
\usage{
path(...)
}
\arguments{
\item{...}{a sequence of quoted or unquoted character strings specifying
JSON object names}
}
\value{
a \code{path} object
}
\description{
Create a JSON path with a minimum of typing
}
\keyword{internal}
