% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbind_tbl_json}
\alias{rbind_tbl_json}
\title{Bind two tbl_json objects together and preserve JSON attribute}
\usage{
rbind_tbl_json(x, y)
}
\description{
Bind two tbl_json objects together and preserve JSON attribute
}

