% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_json.r
\name{read_json}
\alias{read_json}
\title{Reads JSON from an input uri (file, url, ...) and returns a tbl_json}
\usage{
read_json(path, json.lines = "infer")
}
\arguments{
\item{path}{to some json data}

\item{json.lines}{If True, process the data one JSON record per line.
If False, process the data like one large JSON record.
If infer, evaluate True only if the suffix of the filepath ends in .jsonl}
}
\value{
tbl_json instance
}
\description{
Reads JSON from an input uri (file, url, ...) and returns a tbl_json
}

