% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/path.r
\name{replace_nulls}
\alias{replace_nulls}
\title{Replace nulls with something else}
\usage{
replace_nulls(l, replace)
}
\arguments{
\item{l}{a list}

\item{replace}{what to replace NULLs in the list with}
}
\description{
Replace nulls with something else
}

