% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_all.R
\name{spread_all}
\alias{spread_all}
\title{Spreads all non-array keys into new columns}
\usage{
spread_all(.x, sep = ".")
}
\arguments{
\item{.x}{a json string or tbl_json object}

\item{sep}{what to separate nested object keys with}
}
\description{
Spreads all non-array keys into new columns
}

