% Generated by roxygen2 (4.1.0): do not edit by hand
\name{spread_values}
\alias{spread_values}
\title{Extracts values from JSON refereced by a sequence of keys}
\usage{
spread_values(x, ...)
}
\arguments{
\item{x}{tbl_json object}

\item{...}{column=value list where 'column' will be the column name created
and 'value' must be a call to jstring(), jnumber() or jlogical() specifying
the path to get the value (and the type implicit in the function name)}
}
\description{
Extracts values from JSON refereced by a sequence of keys
}

