% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tbl_json.r
\name{[.tbl_json}
\alias{[.tbl_json}
\title{Extract subsets of a tbl_json object (not replace)}
\usage{
\method{[}{tbl_json}(x, i, j, drop = if (missing(i)) TRUE else length(cols) ==
  1)
}
\arguments{
\item{x}{a tbl_json object}

\item{i}{row elements to extract}

\item{j}{column elements to extract}

\item{drop}{whether or not to simplify results}
}
\description{
Extends `[.data.frame` to work with tbl_json objects, so that row filtering
of the underlying data.frame also filters the associated JSON.
}

