% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_df}
\alias{as_data_frame}
\alias{tbl_df}
\title{Convert object to tbl_df}
\usage{
tbl_df(data)
}
\description{
Exported from dplyr package.  Converts an object
to a tbl_df.
}
\seealso{
as_data_frame.tbl_json
}
\keyword{internal}

