% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tbl_json}
\alias{as.tbl_json}
\alias{as.tbl_json.character}
\alias{as.tbl_json.tbl_json}
\alias{is.tbl_json}
\alias{tbl_json}
\title{Combines structured JSON (as a data.frame) with remaining JSON}
\usage{
tbl_json(df, json.list)

as.tbl_json(x, ...)

\method{as.tbl_json}{tbl_json}(x, ...)

is.tbl_json(x)

\method{as.tbl_json}{character}(x, ...)
}
\arguments{
\item{df}{data.frame}

\item{json.list}{list of json lists parsed with fromJSON}

\item{x}{character vector of json}
}
\description{
Combines structured JSON (as a data.frame) with remaining JSON

Note that json.list must have the same length as nrow(df), and if json.list
has any NULL elements, the corresponding rows will be removed from df.

Turns a character vector into a tbl_json object
}

