% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tbl_json.r
\name{wrap_dplyr_verb}
\alias{wrap_dplyr_verb}
\title{Wrapper for extending dplyr verbs to tbl_json objects}
\usage{
wrap_dplyr_verb(dplyr.verb)
}
\arguments{
\item{dplyr.verb}{a dplyr::verb such as filter, arrange}
}
\description{
Wrapper for extending dplyr verbs to tbl_json objects
}

