% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_types.R
\docType{data}
\name{allowed_json_types}
\alias{allowed_json_types}
\title{Fundamental JSON types from http://json.org/, where I collapse 'true' and
'false' into 'logical'}
\format{
An object of class \code{character} of length 6.
}
\usage{
allowed_json_types
}
\description{
Fundamental JSON types from http://json.org/, where I collapse 'true' and
'false' into 'logical'
}
\keyword{datasets}
