% Generated by roxygen2 (4.0.1): do not edit by hand
\name{append_values}
\alias{append_values}
\alias{append_values_logical}
\alias{append_values_number}
\alias{append_values_string}
\title{Append keys to a new column}
\usage{
append_values_string(x, column.name = type)

append_values_number(x, column.name = type)

append_values_logical(x, column.name = type)
}
\arguments{
\item{x}{a tbl_json object}

\item{column.name}{the column.name to append the values into the data.frame
under}
}
\description{
Append keys to a new column
}

