% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_values.R
\name{append_values_factory}
\alias{append_values_factory}
\title{Creates the append_values_* functions}
\usage{
append_values_factory(type, as.value)
}
\arguments{
\item{type}{the JSON type that will be appended}

\item{as.value}{function to force coercion to numeric, string, or logical}
}
\description{
Creates the append_values_* functions
}
\keyword{internal}
