% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_values.R
\name{append_values_type}
\alias{append_values_type}
\title{get list of values from json}
\usage{
append_values_type(json, type)
}
\arguments{
\item{json}{extracted using attributes}

\item{type}{input type (numeric, string, etc)}
}
\description{
get list of values from json
}
\keyword{internal}
