% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{as_tibble}
\alias{as_tibble}
\alias{as_data_frame}
\alias{tbl_df}
\alias{reexports}
\title{Objects exported from other packages}
\details{
Exported from dplyr package.  Converts an object
to a pure tibble (revert to tbl_df class and drops
tbl_json class/attributes).
}
\seealso{
as_tibble.tbl_json
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{as_data_frame}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

