% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-commits.r
\docType{data}
\name{commits}
\alias{commits}
\title{Commit data for the dplyr repo from github API}
\format{JSON}
\usage{
commits
}
\description{
Commit data for the dplyr repo from github API
}
\examples{
library(dplyr)

# Show first 200 characters of JSON
commits \%>\% substr(1, 200) \%>\% writeLines

# Extract metadata for every commit
commits \%>\%   # single json document of github commits from dplyr
  gather_array \%>\%  # stack as an array
  spread_values(
    sha         = jstring("sha"),
    author      = jstring("commit", "author", "name"),
    author.date = jstring("commit", "author", "date")
  ) \%>\% head
}

