% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gather_array}
\alias{gather_array}
\title{Expands a jdf to span the indices of a JSON array}
\usage{
gather_array(x, column.name = "array.index")
}
\arguments{
\item{x}{a jdf whose JSON attribute should always be an array}

\item{column.name}{the name to give to the array index column created}
}
\value{
a jdf with a new column (column.name) that captures the array
  index and JSON attribute extracted from the array
}
\description{
Expands a jdf to span the indices of a JSON array
}

