% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{issues}
\alias{issues}
\title{Issue data for the dplyr repo from github API}
\format{JSON}
\usage{
issues
}
\description{
Issue data for the dplyr repo from github API
}
\examples{
library(dplyr)

# Show first 2k characters of JSON
issues \%>\% substr(1, 2000) \%>\% writeLines

# Extract metadata for every issue
issues \%>\%          # single json document of github issues from dplyr
  as.jdf \%>\%        # turn into a 'jdf'
  gather_array \%>\%  # stack as an array
  spread_values(
    id          = jnumber("id"),
    number      = jnumber("number"),
    title       = jstring("title"),
    user.login  = jstring("user", "login"),
    sate        = jstring("state"),
    locked      = jlogical("locked"),
    comments    = jnumber("comments")
  )

# Extract label content for issues with labels
issues \%>\%          # single json document of github issues from dplyr
  as.jdf \%>\%        # turn into a 'jdf'
  gather_array \%>\%  # stack as an array
  spread_values(id = jnumber("id")) \%>\% # capture issue id for relational purposes
  enter_object("labels") \%>\%            # filter just those with labels
  gather_array("label.index") \%>\%       # stack labels
  spread_values(
    url   = jstring("url"),
    name  = jstring("name"),
    color = jstring("color")
  )
}

