% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jdf}
\alias{as.jdf}
\alias{as.jdf.character}
\alias{as.jdf.jdf}
\alias{is.jdf}
\alias{jdf}
\title{Combines structured JSON (as a data.frame) with remaining JSON}
\usage{
as.jdf(x, ...)

\method{as.jdf}{jdf}(x, ...)

is.jdf(x)

\method{as.jdf}{character}(x, ...)
}
\arguments{
\item{x}{character vector of json}
}
\description{
Combines structured JSON (as a data.frame) with remaining JSON

Turns a character vector into a jdf object
}
\examples{
json <- as.jdf(commits)
attr(json, "JSON")
}

