% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jdf}
\alias{as.jdf}
\alias{as.jdf.character}
\alias{as.jdf.jdf}
\alias{jdf}
\title{jdf class combines structured JSON (as a data.frame) with remaining
parsed JSON}
\usage{
as.jdf(x, ...)

\method{as.jdf}{jdf}(x, ...)

\method{as.jdf}{character}(x, ...)
}
\arguments{
\item{x}{character vector of json}
}
\description{
jdf class combines structured JSON (as a data.frame) with remaining
parsed JSON

Turns a character vector into a jdf object
}

