% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_values.r
\name{jfactory}
\alias{jfactory}
\title{Factory that creates the j* functions below}
\usage{
jfactory(na.value, conversion.function, map.function)
}
\arguments{
\item{na.value}{value to replace NULL with}

\item{conversion.function}{function to convert vector to appropriate type}

\item{map.function}{function to map to collapse}
}
\description{
Factory that creates the j* functions below
}

