% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_values.R
\name{jfactory}
\alias{jfactory}
\title{Factory that creates the j* functions below}
\usage{
jfactory(map.function)
}
\arguments{
\item{map.function}{function to map to collapse}
}
\description{
Factory that creates the j* functions below
}

