% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spread_values.r
\name{jfactory}
\alias{jfactory}
\title{Factory that creates the j* functions below}
\usage{
jfactory(na.value, conversion.function)
}
\arguments{
\item{na.value}{value to replace NULL with}

\item{conversion.function}{function to convert vector to appropriate type}
}
\description{
Factory that creates the j* functions below
}

