% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jfunctions}
\alias{jfunctions}
\alias{jlogical}
\alias{jnumber}
\alias{jstring}
\title{Navigates nested objects to get at keys of a specific type, to be used as
arguments to spread_values}
\usage{
jstring(...)

jnumber(...)

jlogical(...)
}
\arguments{
\item{...}{the path to follow}
}
\value{
a function that can operate on parsed JSON data
}
\description{
Navigates nested objects to get at keys of a specific type, to be used as
arguments to spread_values

Specifies a path to extract for strings

Specifies a path to extract for numbers

Specifies a path to extract for logicals
}

