% Generated by roxygen2 (4.0.1): do not edit by hand
\name{json_types}
\alias{json_types}
\title{Add a column that tells the 'type' of the data in the root of the JSON}
\usage{
json_types(x, column.name = "type")
}
\arguments{
\item{x}{a tbl_json object}

\item{column.name}{the name to specify for the type column}
}
\value{
a tbl_json object with column.name column that tells the type
}
\description{
Add a column that tells the 'type' of the data in the root of the JSON
}

