% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/append_values.r
\name{my_unlist}
\alias{my_unlist}
\title{Unlists while preserving NULLs and only unlisting lists with one value}
\usage{
my_unlist(l)
}
\arguments{
\item{l}{a list that we want to unlist}
}
\description{
Unlists while preserving NULLs and only unlisting lists with one value
}

