% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_values.R
\name{my_unlist}
\alias{my_unlist}
\title{Unlists while preserving NULLs and only unlisting lists with one value}
\usage{
my_unlist(l, recursive = FALSE)
}
\arguments{
\item{l}{a list that we want to unlist}

\item{recursive}{logical indicating whether to unlist nested lists}
}
\description{
Unlists while preserving NULLs and only unlisting lists with one value
}
\keyword{internal}
