% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_json.R
\name{print.tbl_json}
\alias{print.tbl_json}
\title{Print a tbl_json object}
\usage{
\method{print}{tbl_json}(x, ..., json.n = 20, json.width = 15)
}
\arguments{
\item{x}{a \code{\link{tbl_json}} object}

\item{...}{other arguments into \code{\link[tibble]{print.tbl_df}}}

\item{json.n}{number of json records to add (...) otherwise}

\item{json.width}{number of json characters to print}
}
\description{
Print a tbl_json object
}
