% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbind_tbl_json}
\alias{rbind_tbl_json}
\title{Bind two tbl_json objects together and preserve JSON attribute}
\usage{
rbind_tbl_json(x, y)
}
\arguments{
\item{x}{a tbl_json object}

\item{y}{a tbl_json_object}
}
\value{
x and y row-binded together with appropriate JSON attribute
}
\description{
Bind two tbl_json objects together and preserve JSON attribute
}
