% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json.R
\name{read_json}
\alias{read_json}
\title{Reads JSON from an input uri (file, url, ...) and returns a tbl_json}
\usage{
read_json(path, format = c("json", "jsonl", "infer"))
}
\arguments{
\item{path}{to some json data}

\item{format}{If "json", process the data like one large JSON record.
If "jsonl", process the data one JSON record per line (json lines format)
If "infer", the format is the suffix of the given filepath.}
}
\value{
a \code{\link{tbl_json}} object
}
\description{
Reads JSON from an input uri (file, url, ...) and returns a tbl_json
}

