% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_json.R
\name{[.tbl_json}
\alias{[.tbl_json}
\title{Extract subsets of a tbl_json object (not replace)}
\usage{
\method{[}{tbl_json}(.x, i, j, drop = FALSE)
}
\arguments{
\item{.x}{a tbl_json object}

\item{i}{row elements to extract}

\item{j}{column elements to extract}

\item{drop}{whether or not to simplify results}
}
\value{
a \code{\link{tbl_json}} object
}
\description{
Extends `[.data.frame` to work with tbl_json objects, so that row filtering
of the underlying data.frame also filters the associated JSON.
}
